Add files via upload
Girls, If you could try to add the template to your parts, it would be great. it can be without the text I put before the picture. but you can keep the bottom part as well, with the buttons.  I will put everything together at the end, to make sure it works. I have put the CSS as well, if you like it, you could use it too. Let me know if you want me to make any changes